Refactor JSP payloads
